ingestion/requirements.txt
# PDF Processing
PyMuPDF==1.23.8
pdfplumber==0.10.3

# Vector Database & Embeddings
pinecone-client==3.0.0
sentence-transformers==2.7.0

# LLM Integration
langchain==0.1.0
langchain-community==0.0.10

# Utilities
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.4
tqdm==4.66.1

